backend/handler_service/etl/requirements.txt
backoff==2.2.1
psycopg2-binary==2.9.5
pydantic==1.10.7
typing-extensions==4.5.0
certifi==2022.12.7
elastic-transport==8.4.0
elasticsearch==8.6.2
urllib3==1.26.15
orjson==3.8.8
